pyproject.toml
[tool.poetry]
name = "swift-cart"
version = "0.1.0"
description = "Online shop"
authors = ["D4CHub <akbievalim0578@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "5.0.4"
asgiref = "3.8.1"
attrs = "23.2.0"
certifi = "2024.2.2"
charset-normalizer = "3.3.2"
coreapi = "2.3.3"
coreschema = "0.0.4"
django-rest-swagger = "2.2.0"
djangorestframework = "3.15.1"
drf-spectacular = "0.27.0"
idna = "3.7"
inflection = "0.5.1"
itypes = "1.2.0"
jinja2 = "3.1.3"
jsonschema = "4.22.0"
jsonschema-specifications = "2023.12.1"
markupsafe = "2.1.5"
openapi-codec = "1.3.2"
packaging = "24.0"
psycopg2-binary = "2.9.9"
python-dotenv = "1.0.1"
pytz = "2024.1"
pyyaml = "6.0.1"
referencing = "0.35.1"
requests = "2.31.0"
rpds-py = "0.18.0"
simplejson = "3.19.2"
sqlparse = "0.5.0"
uritemplate = "4.1.1"
urllib3 = "2.2.1"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
package-mode = false